% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pond_data.R
\docType{data}
\name{ponds}
\alias{ponds}
\title{Invertebrates collected in temporary ponds}
\format{
An object of class data.frame
}
\usage{
data(ponds)
}
\description{
The test data consist of invertebrate taxa collected in six temporary ponds
in a nature reserve in Belgium. In addition, the data consist of
latitude-longitude coordinates for the samples as well as measured body
lengths for all the taxa.
}
\examples{
data(ponds)

}
\references{
Vincent Dolmans (Vincent.Dolmans@vub.be)
}
\keyword{datasets}
