% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_equation_database.R
\name{view_equation_database}
\alias{view_equation_database}
\title{view_equation_database()}
\usage{
view_equation_database()
}
\value{
tibble with chosen traits or equations based on the input parameters
}
\description{
Get the equation database as a tibble()
}
\details{
This function loads the equation database as a tibble() so that
users can examine and filter if if they want to find relevant equations.
}
\author{
James G. Hagan (james_hagan(at)outlook.com)
}
